requirements.txt
fastapi==0.103.0
Pillow==10.0.0
Requests==2.31.0
transformers==4.31.0
python-multipart
uvicorn
python-dotenv
einops
chardet
charset_normalizer
